cleanup major: parameterize current-input/output-port
this makes things _a lot_ simpler! I don't know why this wasn't done
before. possibly because it was hard to debug with current-ports set
so I couldn't log properly.

this does change the nrepl-loop API, but that was never documented
anyhow.

what's nice about this is that spawn! is now called with
current-input/output-port bound to the TCP connection. we can thus
use (print ...) as usual. Also, it allows `nrepl-loop` to be used in a
context where current-error-port might be something other than
current-output-port (over TCP, you can't distinguish). this might come
in handy if we ever provide nrepl over SSH!